+++ src/superswitch18.cpp
@@ -4,6 +4,7 @@
 #include "sanguinehelpers.hpp"
 #include "pcg_variants.h"
 #include "switches.hpp"
+#include "manus.hpp"
 
 using simd::float_4;
 
@@ -51,6 +52,11 @@ struct SuperSwitch18 : SanguineModule {
 		OUTPUTS_COUNT
 	};
 
+	enum LightIds {
+		LIGHT_EXPANDER,
+		LIGHTS_COUNT
+	};
+
 	dsp::BooleanTrigger btDecrease;
 	dsp::BooleanTrigger btIncrease;
 	dsp::BooleanTrigger btRandom;
@@ -60,6 +66,7 @@ struct SuperSwitch18 : SanguineModule {
 	dsp::SchmittTrigger stInputIncrease;
 	dsp::SchmittTrigger stInputRandom;
 	dsp::SchmittTrigger stInputReset;
+	dsp::SchmittTrigger stDirectSteps[kMaxSteps];
 	bool bClockReceived = false;
 	bool bLastOneShotValue = false;
 	bool bLastResetToFirstStepValue = true;
@@ -67,11 +74,14 @@ struct SuperSwitch18 : SanguineModule {
 	bool bOneShot = false;
 	bool bOneShotDone = false;
 	bool bResetToFirstStep = true;
+	bool bHasExpander = false;
 	int channelCount = 0;
 	int randomNum = 0;
 	int selectedOut = 0;
 	int stepCount = kMaxSteps;
 	int stepsDone = 0;
+
+	static const int kClockDivider = 16;
 
 	float_4 outVoltages[4] = {};
 
@@ -80,7 +90,7 @@ struct SuperSwitch18 : SanguineModule {
 	pcg32_random_t pcgRng;
 
 	SuperSwitch18() {
-		config(PARAMS_COUNT, INPUTS_COUNT, OUTPUTS_COUNT, 0);
+		config(PARAMS_COUNT, INPUTS_COUNT, OUTPUTS_COUNT, LIGHTS_COUNT);
 
 		configParam(PARAM_STEPS, 1.f, 8.f, 8.f, "Steps", "", 0.f, 1.f, 0.f);
 		paramQuantities[PARAM_STEPS]->snapEnabled = true;
@@ -109,7 +119,7 @@ struct SuperSwitch18 : SanguineModule {
 		params[PARAM_STEP1].setValue(1);
 		params[PARAM_RESET_TO_FIRST_STEP].setValue(1);
 		pcg32_srandom_r(&pcgRng, std::round(system::getUnixTime()), (intptr_t)&pcgRng);
-		clockDivider.setDivision(16);
+		clockDivider.setDivision(kClockDivider);
 	};
 
 	void doDecreaseTrigger() {
@@ -180,7 +190,13 @@ struct SuperSwitch18 : SanguineModule {
 	};
 
 	void process(const ProcessArgs& args) override {
+		Module* manusExpander = getRightExpander().module;
+
+		bHasExpander = (manusExpander && manusExpander->getModel() == modelManus && !manusExpander->isBypassed());
+
 		if (clockDivider.process()) {
+			const float sampleTime = args.sampleTime * kClockDivider;
+
 			if (inputs[INPUT_STEPS].isConnected()) {
 				stepCount = round(clamp(inputs[INPUT_STEPS].getVoltage(), 1.f, 8.f));
 			} else if (params[PARAM_STEPS].getValue() != stepCount) {
@@ -193,6 +209,19 @@ struct SuperSwitch18 : SanguineModule {
 					params[PARAM_STEP1 + step].setValue(step == selectedOut ? 1 : 0);
 				} else {
 					params[PARAM_STEP1 + step].setValue(2);
+				}
+			}
+
+			lights[LIGHT_EXPANDER].setBrightnessSmooth(bHasExpander ? 0.75f : 0.f, sampleTime);
+
+			if (bHasExpander) {
+				for (int step = 0; step < kMaxSteps; ++step) {
+					int currentLight = Manus::LIGHT_STEP_1_LEFT + step * 3;
+					if (step < stepCount) {
+						manusExpander->getLight(currentLight).setBrightnessSmooth(0.75f, sampleTime);
+					} else {
+						manusExpander->getLight(currentLight).setBrightnessSmooth(0.f, sampleTime);
+					}
 				}
 			}
 		}
@@ -224,6 +253,14 @@ struct SuperSwitch18 : SanguineModule {
 
 			if (bResetToFirstStep || (!bResetToFirstStep && bClockReceived)) {
 				for (int output = 0; output < stepCount; ++output) {
+					if (bHasExpander) {
+						const int currentInput = Manus::INPUT_STEP_1 + output;
+						if (stDirectSteps[output].process(manusExpander->getInput(currentInput).getVoltage()) && output < stepCount) {
+							selectedOut = output;
+						}
+					}
+
+
 					if (btSteps[output].process(params[PARAM_STEP1 + output].getValue()) && output < stepCount) {
 						if (selectedOut > -1) {
 							outputs[OUTPUT_OUT1 + selectedOut].setChannels(0);
@@ -303,6 +340,22 @@ struct SuperSwitch18 : SanguineModule {
 		selectedOut = static_cast<int>(pcg32_boundedrand_r(&pcgRng, stepCount));
 	}
 
+	void onBypass(const BypassEvent& e) override {
+		if (bHasExpander) {
+			Module* manusExpander = getRightExpander().module;
+			manusExpander->getLight(Manus::LIGHT_MASTER_MODULE_LEFT).setBrightness(0.f);
+		}
+		Module::onBypass(e);
+	}
+
+	void onUnBypass(const UnBypassEvent& e) override {
+		if (bHasExpander) {
+			Module* manusExpander = getRightExpander().module;
+			manusExpander->getLight(Manus::LIGHT_MASTER_MODULE_LEFT).setBrightness(0.75f);
+		}
+		Module::onUnBypass(e);
+	}
+
 	json_t* dataToJson() override {
 		json_t* rootJ = SanguineModule::dataToJson();
 
@@ -348,6 +401,8 @@ struct SuperSwitch18Widget : SanguineModuleWidget {
 		makePanel();
 
 		addScrews(SCREW_ALL);
+
+		addChild(createLightCentered<SmallLight<OrangeLight>>(millimetersToPixelsVec(63.44f, 9.672f), module, SuperSwitch18::LIGHT_EXPANDER));
 
 		addParam(createParamCentered<BefacoTinyKnobRed>(millimetersToPixelsVec(9.844, 18.272), module, SuperSwitch18::PARAM_STEPS));
 		addInput(createInputCentered<BananutBlack>(millimetersToPixelsVec(9.844, 32.461), module, SuperSwitch18::INPUT_STEPS));
@@ -445,6 +500,22 @@ struct SuperSwitch18Widget : SanguineModuleWidget {
 
 		SanguineShapedLight* bloodLogo = new SanguineShapedLight(module, "res/blood_glowy_small.svg", 26.154, 112.707);
 		addChild(bloodLogo);
+	}
+
+	void appendContextMenu(Menu* menu) override {
+		SanguineModuleWidget::appendContextMenu(menu);
+
+		SuperSwitch18* superSwitch18 = dynamic_cast<SuperSwitch18*>(this->module);
+
+		menu->addChild(new MenuSeparator());
+		const Module* expander = superSwitch18->rightExpander.module;
+		if (expander && expander->model == modelManus) {
+			menu->addChild(createMenuLabel("Manus expander already connected"));
+		} else {
+			menu->addChild(createMenuItem("Add Manus expander", "", [=]() {
+				superSwitch18->addExpander(modelManus, this);
+				}));
+		}
 	}
 };